# backend/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "GigConnect Backend API Platform",
    "description": "Clean, systematic, and fully categorized RESTful API documentation for GigConnect services, workers, bookings, admin operations, and background workflows.",
    "version": "1.0.0"
  },
  "host": "localhost:8000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Enter JWT Bearer token: Bearer <token>"
    }
  },
  "paths": {
...
    "/api/workers/": {
      "get": {
        "tags": [
          "Workers"
        ],
        "description": "Get nearby available workers with infinite scroll / lazy loading (5 items per batch), category & skills filtering, top rating sort, and Indian currency minimum charge",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "x-device-id",
            "in": "header",
            "type": "string"
          },
          {
            "name": "lng",
            "in": "query",
            "description": "User longitude (e.g. 77.3639)",
            "required": true,
            "type": "number"
          },
          {
            "name": "lat",
            "in": "query",
            "description": "User latitude (e.g. 28.6280)",
            "required": true,
            "type": "number"
          },
          {
            "name": "category",
            "in": "query",
            "description": "Filter by category (e.g. Plumbing)",
            "type": "string"
          },
          {
            "name": "skill",
            "in": "query",
            "description": "Filter by single skill (e.g. Pipe Fitting)",
            "type": "string"
          },
          {
            "name": "skills",
            "in": "query",
            "description": "Filter by multiple skills comma-separated (e.g. Pipe Fitting, Leak Detection)",
            "type": "string"
          },
          {
            "name": "sortBy",
            "in": "query",
            "description": "Sorting order: top_rated (default), nearest, jobs, price_low, price_high",
            "type": "string",
            "default": "top_rated"
          },
          {
            "name": "radiusInKm",
            "in": "query",
            "description": "Search radius in kilometers (default 10)",
            "type": "number",
            "default": 10
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Infinite scroll offset (0 for first batch, 5 for next batch, etc.)",
            "type": "integer",
            "default": 0
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Batch size limit from .env (default 5)",
            "type": "integer",
            "default": 5
          },
          {
            "name": "page",
            "in": "query",
            "description": "Alternative page number for scroll (1 for first batch, 2 for next, etc.)",
            "type": "integer",
            "default": 1
          },
          {
            "name": "skip",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "summary": "Find Nearby Available Workers (Geo-query & Filters)"
      }
    },
...
    "/api/services/": {
      "post": {
        "description": "Create New Service (Uses Cloudinary)",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                },
                "category": {
                  "example": "any"
                },
                "basePrice": {
                  "example": "any"
                },
                "estimatedTime": {
                  "example": "any"
                },
                "whatsIncluded": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Services"
        ],
        "summary": "Create New Service (Uses Cloudinary)"
      }
    },
...
    "/api/bookings/": {
      "post": {
        "tags": [
          "Bookings"
        ],
        "description": "Customer creates a new booking request with scheduling and emergency SOS support",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "x-device-id",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "serviceId": {
                  "example": "any"
                },
                "workerId": {
                  "example": "any"
                },
                "problemDescription": {
                  "example": "any"
                },
                "problemPhotos": {
                  "example": "any"
                },
                "addressLine": {
                  "example": "any"
                },
                "coordinates": {
                  "example": "any"
                },
                "scheduledTime": {
                  "example": "any"
                },
                "timeSlot": {
                  "example": "any"
                },
                "bookingType": {
                  "example": "any"
                },
                "isEmergency": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "summary": "Create New Service Booking",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/admin/bookings": {
      "get": {
        "description": "List All Bookings across Platform (Filtered by Status & Date)",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          },
          {
            "name": "search",
            "in": "query",
            "type": "string"
          },
          {
            "name": "status",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Admin"
        ],
        "summary": "List All Bookings across Platform (Filtered by Status & Date)",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/admin/cooperative": {
      "get": {
        "description": "Get Cooperative Governance Overview",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "tags": [
          "Admin"
        ],
        "summary": "Get Cooperative Governance Overview",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "description": "Update Cooperative Governance Policy",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "tags": [
          "Admin"
        ],
        "summary": "Update Cooperative Governance Policy",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/admin/notifications": {
      "get": {
        "description": "List Sent Broadcast Notifications",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Admin"
        ],
        "summary": "List Sent Broadcast Notifications",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/admin/payments": {
      "get": {
        "description": "List All Platform Financial Transactions",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          },
          {
            "name": "status",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Admin"
        ],
        "summary": "List All Platform Financial Transactions",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/admin/reviews": {
      "get": {
        "description": "Moderate All Customer Reviews & Ratings",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          },
          {
            "name": "rating",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Admin"
        ],
        "summary": "Moderate All Customer Reviews & Ratings",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/admin/services": {
      "get": {
        "description": "List All Services & Pricing Catalog",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          },
          {
            "name": "search",
            "in": "query",
            "type": "string"
          },
          {
            "name": "category",
            "in": "query",
            "type": "string"
          },
          {
            "name": "isActive",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Admin"
        ],
        "summary": "List All Services & Pricing Catalog",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "description": "Create New Service with Image Upload & Redis Push",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "tags": [
          "Admin"
        ],
        "summary": "Create New Service with Image Upload & Redis Push",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/admin/upload": {
      "post": {
        "description": "Upload Single Media File to Cloudinary",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Admin"
        ],
        "summary": "Upload Single Media File to Cloudinary",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/admin/workers": {
      "get": {
        "description": "List All Workers with Verification & Category Filters",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          },
          {
            "name": "search",
            "in": "query",
            "type": "string"
          },
          {
            "name": "category",
            "in": "query",
            "type": "string"
          },
          {
            "name": "isVerified",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Admin"
        ],
        "summary": "List All Workers with Verification & Category Filters",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "description": "Register New Worker Manually by Admin",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "phone": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "category": {
                  "example": "any"
                },
                "rate": {
                  "example": "any"
                },
                "experienceYears": {
                  "example": "any"
                },
                "bio": {
                  "example": "any"
                },
                "skills": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Admin"
        ],
        "summary": "Register New Worker Manually by Admin",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/payments/worker-wallet": {
      "get": {
        "description": "Get Worker Wallet Financial Balance",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "x-device-id",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Payments"
        ],
        "summary": "Get Worker Wallet Financial Balance"
      }
    },
...
    "/api/home/home": {
      "get": {
        "description": "Get Home Dashboard Data (Banners, Categories & Top Services)",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "x-device-id",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Home"
        ],
        "summary": "Get Home Dashboard Data (Banners, Categories & Top Services)"
      }
    },
...
    "/api/worker-certificates/": {
      "post": {
        "description": "Upload New Skill Certificate for Verification",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "serviceCategory": {
                  "example": "any"
                },
                "certificateType": {
                  "example": "any"
                },
                "certificateNumber": {
                  "example": "any"
                },
                "fileUrl": {
                  "example": "any"
                },
                "issuer": {
                  "example": "any"
                },
                "issuedOn": {
                  "example": "any"
                },
                "expiresOn": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "KYC & Verification"
        ],
        "summary": "Upload New Skill Certificate for Verification"
      },
      "get": {
        "description": "List Worker Certificates",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "KYC & Verification"
        ],
        "summary": "List Worker Certificates"
      }
    },
...
    "/api/notifications/": {
      "get": {
        "description": "List User In-App Notifications",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Notifications"
        ],
        "summary": "List User In-App Notifications"
      }
    },
...
    "/api/welfare/": {
      "get": {
        "description": "Get Worker Welfare Program Overview & Fund Status",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Welfare"
        ],
        "summary": "Get Worker Welfare Program Overview & Fund Status"
      }
    },
...
    "/api/upload/": {
      "post": {
        "description": "Upload Single Image / Document to Cloudinary",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "x-device-id",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Upload"
        ],
        "summary": "Upload Single Image / Document to Cloudinary"
      }
    },
...
  },
...
}